GO-1069 - Introduce "promotion" service and integrate with exisiting AUTO_OPT_IN flow

+++ shared/src/main/java/mobi/nowtechnologies/server/shared/Utils.java
@@ -105,7 +105,7 @@ public class Utils {
 		return new Date(((long) intDate) * 1000L);
 	}
 
-    public static long secondsToMillis(int seconds){
+    public static long secondsToMillis(long seconds){
         return SECONDS.toMillis(seconds);
     }
